Add more preview for MediaDetailsBottomSheetPreview
